+++ CHANGELOG.md
@@ -9,8 +9,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `scalr_workspace`: new attribute `terragrunt` ([#390](https://github.com/Scalr/terraform-provider-scalr/pull/390))
+ - **New resource:** `scalr_workload_identity_provider` ([#405](https://github.com/Scalr/terraform-provider-scalr/pull/405)
+ - **New resource:** `scalr_assume_service_account_policy` ([#405](https://github.com/Scalr/terraform-provider-scalr/pull/405)
+
+### Changed
+
+- `scalr_environment`: attribute `policy_groups` became read-only ([#395](https://github.com/Scalr/terraform-provider-scalr/pull/395))
+
+## [2.6.0] - 2025-03-07
+
+### Added
+
 - `scalr_environment` and `data.scalr_environment`: new attribute `mask_sensitive_output` ([#389](https://github.com/Scalr/terraform-provider-scalr/pull/389))
+- **New resource:** `scalr_integration_infracost` ([#397](https://github.com/Scalr/terraform-provider-scalr/pull/397)
+- **New data source:** `scalr_integration_infracost` ([#397](https://github.com/Scalr/terraform-provider-scalr/pull/397))
+
+### Fixed
+
+- `data.scalr_provider_configuration` not populating the `name` attribute ([#399](https://github.com/Scalr/terraform-provider-scalr/pull/399))
+- `scalr_workspace`: issue with `run_operation_timeout` attribute causing "provider produced an unexpected new value" error on resource update ([#401](https://github.com/Scalr/terraform-provider-scalr/pull/401))
+
+### Required
+
+- scalr-server >= 8.177.0
+
+## [2.5.0] - 2025-02-07
+
+### Added
+
+- `scalr_workspace`: new attribute `terragrunt` ([#390](https://github.com/Scalr/terraform-provider-scalr/pull/390))
 
 ### Changed
 
@@ -20,11 +47,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `scalr_workspace`: deleted attribute `terragrunt_version` ([#390](https://github.com/Scalr/terraform-provider-scalr/pull/390))
 - `scalr_workspace`: deleted attribute `terragrunt_use_run_all` ([#390](https://github.com/Scalr/terraform-provider-scalr/pull/390))
-- `scalr_environment`: attribute `policy_groups` became read-only ([#395](https://github.com/Scalr/terraform-provider-scalr/pull/395))
 
 ### Required
 
-- scalr-server >= 8.172.0
+- scalr-server >= 8.174.0
 
 ## [2.4.0] - 2025-01-24
 
@@ -943,7 +969,9 @@ Requires Scalr 8.0.1-beta.20200625 at least
 
 - Initial release.
 
-[Unreleased]: https://github.com/Scalr/terraform-provider-scalr/compare/v2.4.0...HEAD
+[Unreleased]: https://github.com/Scalr/terraform-provider-scalr/compare/v2.6.0...HEAD
+[2.6.0]: https://github.com/Scalr/terraform-provider-scalr/releases/tag/v2.6.0
+[2.5.0]: https://github.com/Scalr/terraform-provider-scalr/releases/tag/v2.5.0
 [2.4.0]: https://github.com/Scalr/terraform-provider-scalr/releases/tag/v2.4.0
 [2.3.0]: https://github.com/Scalr/terraform-provider-scalr/releases/tag/v2.3.0
 [2.2.0]: https://github.com/Scalr/terraform-provider-scalr/releases/tag/v2.2.0